\name{cairns.foster.data}
\alias{cairns.foster.data}
\docType{data}
\title{Experimental Data from Cairns and Foster}

\description{Well-known data, analyzed by Rosche and Foster (2000).}

\details{
Experiment described in detail in Cairns and Foster (1991), data first appeared in
Rosche and Foster (2000) with the maximum likelihood estimate of m explicitly given.
Note Nt=1.5e8.
}

\source{
WA Rosche and PL Foster, Determining mutation rates in bacterial population, 
Methods 20: 4-17 (2000)
}

\references{
J Cairns and PL Foster, Adaptive reversion of a frameshift mutation in Escherichia coli,
Genetics 128:695-701 (1991)
}
\examples{
newton.LD(cairns.foster.data,show.iter=TRUE)
}
