\name{foster.data}
\alias{foster.data}
\docType{data}
\title{
Fluctuation Experimental Data by P. Foster
}
\description{
Data from an experiment conducted by P. Foster in 1994. The study was about a reversion of Lac- to Lac+
in the FC 40 strain of Escherichia coli. Three cultures were sampled to to estimate Nt. At the 1e-6
dilution, the three measurements were 702, 645 and 536. Therefore, Nt=6.16e8. The data first appeared
in Zheng (2011).
}
\references{
Q. Zheng, A Bayesian two-level model for fluctuation assay,
Genetica 139 (2011) 1409-1416.
}
\examples{
newton.LD(foster.data)
}
\keyword{datasets}
