\name{log.likelihood.LD.plating}
\alias{log.likelihood.LD.plating}

\title{Computing the Log-Likelihood Function That Adjusts for Plating Efficiency} 

\description{This function computes the log-likelihood function that adjusts for plating
efficiency. The parameter m is regard as the argument of the log-likelihood function.}

\usage{log.likelihood.LD.plating(data, m, e)}

\arguments{
  \item{data}{A vector of experimental data.}
  \item{m}{A real positive number as a given value of the parameter m.}
  \item{e}{plating efficiency.}
}

\author{Qi Zheng <qzheng@sph.tamhsc.edu>}

\seealso{log.likelihood.LD}

\examples{}
