\name{luria.16.data}
\alias{luria.16.data}
\docType{data}
\title{Classic Experimental Data from Luria and Delbruck}

\description{Data taken from Luria and Delbruck (1943) and analyzed in Zheng (2008).}

\details{In this experiment, each test tube had a 0.2-ml culture, but only a portion of 0.08 ml
was plated. Therefore, the plating efficiency was e=0.4.}


\references{
S.E. Luria, M. Delbruck, Mutations of bacteria
from virus sensitivity to virus resistance, Genetics 28 (1943) 491-511.

Q. Zheng. A note on plating efficiency in fluctuation experiments,
Mathematical Biosciences, 216 (2008) 150-153.
}

\examples{newton.LD.plating(luria.16.data, e=0.4, show.iter = TRUE)}
