\name{niccum.data}
\alias{niccum.data}
\docType{data}
\title{An Experiment Reported by Niccum et al.}

\description{Data taken from Niccum et al. (2012).}

\details{This data set is the Time 1 part of Experiment A of Niccum et al. (2012).
The estimated number of viable cells is 31350000, close to 33554432. As N0=1 was assumed,
the number of generations is therefore g=25.}


\references{
Niccum, BA, Poteau, R, Hamman, GE, Varada, JC, Dshalalow, JH, Sinden, RR, 2012. On
an unbiased and consistent estimator for mutation rates, Journal of Theoretical Biology 300,
360-367.
}

\examples{newton.Haldane(niccum.data,g=25)}

