\name{rsalvador-package}
\alias{rsalvador-package}
\alias{rsalvador}
\docType{package}
\title{rSalvador: An R Tool for the Luria-Delbruck Fluctuation Assay.}
\description{Adapted from SALVADOR 2.3 (Zheng, 2008)}

\details{
\tabular{ll}{
Package: \tab rSalvador\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2015-06-16\cr
License: \tab GPL 2.0  \cr
}
}

\author{Qi Zheng <qzheng@sph.tamhsc.edu>}

\references{
Q. Zheng, SALVADOR 2.3: A Tool for Studying Mutation Rates, published in September 2008,
available at http://library.wolfram.com/infocenter/MathSource/7082.
}

\note{Eric H. Zheng provided technical support.}

\keyword{package}
