\name{niccum.data}
\alias{niccum.data}
\docType{data}
\title{An Experiment Reported by Niccum et al.}

\description{Data reported by Niccum et al. (2012).}

\details{This data set is the Time 1 part of Experiment A of Niccum et al. (2012).
The estimated number of viable cells is 31350000 and the plating efficiency is 0.10.}


\references{
BA Niccum, R  Poteau, GE Hamman, JC Varada, JH  Dshalalow, RR Sinden, On an unbiased
and consistent estimator for mutation rates, Journal of Theoretical Biology 300 
(2012) 360-367.
}

\examples{newton.LD.plating(niccum.data,e=0.1)}

