\name{rsalvador-package}
\alias{rsalvador-package}
\alias{rsalvador}
\docType{package}
\title{rSalvador: An R Tool for the Luria-Delbruck Fluctuation Assay.}
\description{Adapted from SALVADOR 2.3 (Zheng, 2008). An overview of the package's
functionalities is given in Zheng (2015).}

\details{
\tabular{ll}{
Package: \tab rSalvador\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2015-10-16\cr
License: \tab GPL 2.0  \cr
}
}

\author{Qi Zheng <qzheng@sph.tamhsc.edu>}

\references{
Q. Zheng, SALVADOR 2.3: A Tool for Studying Mutation Rates, published in September 2008,
available at http://library.wolfram.com/infocenter/MathSource/7082.

Q. Zheng, A new practical guide to the Luria-Delbruck protocol,  Mutation
Research/Fundamental and Molecular Mechanisms of Metagenesis 781 (2015) 7-13.



}

\note{Eric H. Zheng provided technical support.}

\keyword{package}
